+++ src/Traits/NamespaceTrait.php
@@ -24,11 +24,13 @@ use Elastic\Elasticsearch\Endpoints\Cluster;
 use Elastic\Elasticsearch\Endpoints\DanglingIndices;
 use Elastic\Elasticsearch\Endpoints\Enrich;
 use Elastic\Elasticsearch\Endpoints\Eql;
+use Elastic\Elasticsearch\Endpoints\Esql;
 use Elastic\Elasticsearch\Endpoints\Features;
 use Elastic\Elasticsearch\Endpoints\Fleet;
 use Elastic\Elasticsearch\Endpoints\Graph;
 use Elastic\Elasticsearch\Endpoints\Ilm;
 use Elastic\Elasticsearch\Endpoints\Indices;
+use Elastic\Elasticsearch\Endpoints\Inference;
 use Elastic\Elasticsearch\Endpoints\Ingest;
 use Elastic\Elasticsearch\Endpoints\License;
 use Elastic\Elasticsearch\Endpoints\Logstash;
@@ -134,6 +136,15 @@ trait NamespaceTrait
 	}
 
 
+	public function esql(): Esql
+	{
+		if (!isset($this->namespace['Esql'])) {
+			$this->namespace['Esql'] = new Esql($this);
+		}
+		return $this->namespace['Esql'];
+	}
+
+
 	public function features(): Features
 	{
 		if (!isset($this->namespace['Features'])) {
@@ -176,6 +187,15 @@ trait NamespaceTrait
 			$this->namespace['Indices'] = new Indices($this);
 		}
 		return $this->namespace['Indices'];
+	}
+
+
+	public function inference(): Inference
+	{
+		if (!isset($this->namespace['Inference'])) {
+			$this->namespace['Inference'] = new Inference($this);
+		}
+		return $this->namespace['Inference'];
 	}